Updated JobManager to allow users to register completion callbacks that will be called when a job finishes running.
Update OpkgJob to accept a single package or a string of package names to install/uninstall

Updated nmap and tcpdump modules to take advantage of JobManager callbacks

Delete the notifications package in favor of having it be a regular helper function.

Added a send_notification method to module.py that automatically populates the module name in the send notification call.